LayoutTests/crypto/subtle/hmac-export-key-expected.txt

Test exporting an AES key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a JWK key...
Exporting the key as raw data...
PASS bytesToHexString(new Uint8Array(exportedData)) is '6a18e49feff7f3b7e09ec89b7f6deab2f6a18e49feff7f3b7e09ec89b7f6deab'
PASS successfullyParsed is true

TEST COMPLETE
